rename prisma-client to client
