+++ pages/api/auth/[...nextauth].ts
@@ -1,0 +1,4 @@
+import NextAuth from 'next-auth'
+import GoogleProvider from 'next-auth/providers/google'
+import { PrismaAdapter } from '@next-auth/prisma-adapter'
+import prisma from '../../../prisma/prisma-client'
